+++ webpack.config.js
@@ -36,7 +36,8 @@ module.exports = {
     },
     plugins: [
         new HtmlWebpackPlugin({
-            template: 'index.html'
+            template: 'index.html',
+            base: "/"
         }),
         new MiniCssExtractPlugin()
     ]
